lb: moving maglev to extensions
