+++ source/extensions/load_balancing_policies/maglev/maglev_lb.cc
@@ -1,4 +1,4 @@
-#include "source/common/upstream/maglev_lb.h"
+#include "source/extensions/load_balancing_policies/maglev/maglev_lb.h"
 
 #include "envoy/config/cluster/v3/cluster.pb.h"
