+++ test/mocks/upstream/cluster_info.h
@@ -252,6 +252,7 @@ public:
   mutable Http::Http1::CodecStats::AtomicPtr http1_codec_stats_;
   mutable Http::Http2::CodecStats::AtomicPtr http2_codec_stats_;
   mutable Http::Http3::CodecStats::AtomicPtr http3_codec_stats_;
+  ProtobufTypes::MessagePtr load_balancing_policy_;
 };
 
 class MockIdleTimeEnabledClusterInfo : public MockClusterInfo {
